Deferred theme loading for bmpanel2cfg.
Themes are being loaded now in a separate thread. Because the python is
slow and the I/O is slow too it will give a 1-2 second(s) speedup on
the cold file cache startup case.

I think it is always a good idea to minimize stalls on I/O operations.
In these days even some weird scenarios are possible (like FUSE + remote
files location). So.. Let's see how stable this will be.

But generally you shouldn't notice anything. :D