Cleanup the taskbar mess.
1. Instead of using "another monitor" flag, we're using "current
monitor" variable. It is much better, because eventually it will allow
panel to show tasks from multiple certain monitors. So, I should add
this parameter to the config file now.

2. I've added 'is_task_visible' function, it's rationale is described
partially in (1). Since we have a quite complicated condition whether we
need to show the task or not (is task on the current desktop, is task on
all desktops, is task on current monitor, is task on one of monitors
from the visible list, etc.). I think that way it's much better now.