Widget alternatives system.
It allows bmpanel2 to have alternative widgets. For example bmpanel2 is
able to have two different desktop switchers: 'desktop_switcher' and
'pager'. But only one can be used.

The idget alternatives system checks 'preferred_alternatives" config
string for a list of space-separated widget names. This is a list of
what user likes to see on the panel. Only one widget from a group of
conflicting widgets can be choosen as preference.

A decision whether to use widget or not is made using following rules:

1. If user has marked widget as preferred and theme has this widget - it
will be choosen.

2. If user has marked widget as preferred, but theme has no such widget
- the first widget in a theme from the group will be used. It means
that artist can declare his preferred widgets in a theme too. Also this
rule applies if user hasn't marked any widget in this group.

So, it's quite simple.